feat: pie chart
This is a v1 pie chart implementation that renders pie charts, without labels and without gestures.

It sort of only has one option beyond the pie chart, which is to render an inset inside the chart, which is like a border of each slice.

No docs yet either, beside the example page, which only renders in dev mode currently.