fix pie chart/donut chart rendering on mount
it seems like in bumping versions, the polar charts are often blank on first mount, and only visible once some state causes a rerender. it's a bit of a ghost issue, but i think it's better to err on the side of adding a delay to ensure it ends up visible.