update exports values
prefer `react-native` at the top, and point to `src/index` as babel needs to be able to see this code in order to apply reanimated's babel plugin which enables automatic workletization. without this, then the auto worklets aren't applied and we'd need to manually workletize things.

otherwise, defaults to web/node as usual and default points to uncompiled src as well.